# api.yaml
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/{user_id}/rooms": {
      "get": {
        "operationId": "RoomsController_getRooms",
        "summary": "List rooms of a user",
        "description": "Returns all rooms belonging to a user. Optionally filter by apartment (?apartment=apartment_id) and/or apartment location (?location=city).",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartment",
            "required": false,
            "in": "query",
            "description": "Filter by apartment ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "description": "Filter by apartment location (e.g., Oulu)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of rooms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "summary": "Create a new room",
        "description": "Creates a room inside a specific apartment. The apartment ID must be provided via ?apartment=...",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apartment",
            "required": true,
            "in": "query",
            "description": "ID of the apartment",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Room created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "400": {
            "description": "Missing or invalid fields"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch or apartment does not belong to user"
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/user/{user_id}/rooms/{room_id}": {
      "get": {
        "operationId": "RoomsController_getById",
        "summary": "Get a single room by ID",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "room_id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Room not found"
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "RoomsController_update",
        "summary": "Update room details",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "room_id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Room not found"
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RoomsController_delete",
        "summary": "Delete a room",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "room_id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Room deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden – user_id mismatch"
          },
          "404": {
            "description": "Room not found"
          }
        },
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "smart-home",
    "description": "Smart home application",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "room-001",
            "description": "Unique identifier for room"
          },
          "name": {
            "type": "string",
            "example": "Living Room",
            "description": "Name of the room"
          },
          "location": {
            "type": "string",
            "example": "First Floor",
            "description": "Location of the room"
          },
          "apartmentId": {
            "type": "string",
            "example": "apt-111",
            "description": "Unique identifier for apartment assigned"
          }
        },
        "required": [
          "id",
          "name",
          "location",
          "apartmentId"
        ]
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Kitchen",
            "description": "Name of the room"
          },
          "location": {
            "type": "string",
            "example": "Basement",
            "description": "Location of the room"
          }
        },
        "required": [
          "name",
          "location"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Master Bedroom"
          },
          "location": {
            "type": "string",
            "example": "Second Floor"
          }
        }
      },
...
    }
  }
}